New_assignment/requirements.txt
streamlit==1.32.0
pandas==2.2.0
sqlalchemy==2.0.27
pymysql==1.1.0
plotly==5.18.0
scikit-learn==1.4.0
python-dotenv==1.0.1